+++ examples/resources/googleworkspace_user_license/import.sh
@@ -1,0 +1,5 @@
+# Copyright (c) HashiCorp, Inc.
+# SPDX-License-Identifier: MPL-2.0
+
+# User licenses can be imported using the format product_id:sku_id:user_id
+terraform import googleworkspace_user_license.dwight-license Google-Apps:1010020028:dwight.schrute@example.com